refactor(FAQ TAB): faq tab 전용view-modal 제작
